Move some anex function in utils.cpp
