+++ include/utils/utils.h
@@ -3,10 +3,16 @@
 
 #include <string>
 
+#include "indexation/mysql/file.h"
+
 using std::string;
 
 // Returns a random string of length len from a-z characters.
 string random_str(const unsigned int len);
+
+void fillFileRandomly(File *file, bool tweet, bool isEven);
+
+void printSQLException(sql::SQLException &e);
 
 void print(string toPrint, bool verbose);
